requirements.txt
Django==4.2.20
whitenoise==6.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3